refactor(config): make lazy loading proxy abstract
